unibudget-backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
numpy>=1.26.2
pandas>=2.1.0
pydantic>=2.5.2
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
asyncpg>=0.29.0
aiosqlite>=0.19.0
greenlet>=3.0.1
fastapi-cache2>=0.2.1